+++ src/App.vue
@@ -10,21 +10,7 @@
 		</div>
 		<Notifier></Notifier>
 		<Spinner></Spinner>
-		<kb-menu
-			:routing="true"
-			:locale="currentLocale"
-			:page="$route.name"
-		></kb-menu>
-		<div
-			ref="wipe"
-			class="wipe"
-		>
-			<img
-				title="Royal Danish Library"
-				alt="Logo of the Royal Danish Library"
-				:src="getImgServerSrcURL()"
-			/>
-		</div>
+		<Header :locale="currentLocale"></Header>
 	</div>
 	<div class="content">
 		<router-view v-slot="{ Component }">
@@ -39,6 +25,16 @@
 			</transition>
 		</router-view>
 	</div>
+	<div
+		ref="wipe"
+		class="wipe"
+	>
+		<img
+			title="Royal Danish Library"
+			alt="Logo of the Royal Danish Library"
+			:src="getImgServerSrcURL()"
+		/>
+	</div>
 </template>
 
 <script lang="ts">
@@ -50,8 +46,11 @@ import Notifier from '@/components/global/notification/Notifier.vue';
 import Spinner from '@/components/global/spinner/Spinner.vue';
 import { LocalStorageWrapper } from './utils/local-storage-wrapper';
 import Footer from '@/components/global/nav/Footer.vue';
-
+import Header from '@/components/search/Header.vue';
+import { useAuthStore } from '@/store/authStore';
 import '@/components/global/nav/wc-header-menu';
+import { APIService } from '@/api/api-service';
+import { APIAuthMessagesType } from '@/types/APIResponseTypes';
 
 export default defineComponent({
 	name: 'App',
@@ -59,6 +58,7 @@ export default defineComponent({
 		Notifier,
 		Spinner,
 		Footer,
+		Header,
 	},
 	setup() {
 		const td = ref(0.4);
@@ -69,6 +69,7 @@ export default defineComponent({
 		const router = useRouter();
 		const route = useRoute();
 		const { locale, t } = useI18n({ useScope: 'global' });
+		const authStore = useAuthStore();
 
 		const html = document.querySelector('html');
 		html?.setAttribute('lang', 'da');
@@ -145,10 +146,18 @@ export default defineComponent({
 		};
 
 		router.beforeEach((to, from) => {
-			if (from.name === 'Home' && to.name === 'Record') {
+			if (from.name === 'Search' && to.name === 'Record') {
 				transitionName.value = 'from-search-to-record';
+			} else if (from.name === 'Record' && to.name === 'Search') {
+				transitionName.value = 'from-record-to-search';
+			} else if (from.name === 'Home' && to.name === 'Search') {
+				transitionName.value = 'from-record-to-search';
+			} else if (from.name === 'Search' && to.name === 'Home') {
+				transitionName.value = 'from-record-to-search';
 			} else if (from.name === 'Record' && to.name === 'Home') {
 				transitionName.value = 'from-record-to-search';
+			} else if (from.name === 'Home' && to.name === 'Record') {
+				transitionName.value = 'from-search-to-record';
 			} else {
 				transitionName.value = 'swipe';
 			}
@@ -157,6 +166,24 @@ export default defineComponent({
 		onMounted(async () => {
 			//for now, we set the title of the app to the archive. Can be changed if we ever go portal-mode.
 			document.title = t('app.titles.frontpage.archive.name') as string;
+
+			// we try to get the kaltura conf id's here. we got some backup ones from aegis, and they're
+			// fallback if we can't get these. But if the bff backend has some, we use them instead
+
+			Promise.race([APIService.getKalturaConfIds(), new Promise((_, reject) => setTimeout(() => reject(), 5000))])
+				.then((response) => {
+					const typedResponse = response as APIAuthMessagesType; // Assert the correct type
+					authStore.partnerId = typedResponse.data.partnerId;
+					authStore.audioUiConfId = typedResponse.data.AudioUiConfId;
+					authStore.videoUiConfId = typedResponse.data.videoUiConfId;
+					authStore.streamingBaseUrlAudio = typedResponse.data.streamingBaseUrlAudio;
+					authStore.streamingBaseUrlVideo = typedResponse.data.streamingBaseUrlVideo;
+					authStore.kalturaIdFetchExecuted = true;
+				})
+				.catch(() => {
+					authStore.kalturaIdFetchExecuted = true;
+				});
+
 			await router.isReady();
 			const hasLocaleParam = Object.prototype.hasOwnProperty.call(route.query, 'locale');
 			const storedLocale = LocalStorageWrapper.get('locale') as string;
@@ -216,6 +243,10 @@ export default defineComponent({
 
 .result-enter {
 	transition-delay: 0.15s;
+}
+
+.app {
+	position: relative;
 }
 
 .result-enter-from,