+++ src/types/APIResponseTypes.ts
@@ -2,6 +2,7 @@ import { GenericSearchResultType, FacetsType } from './GenericSearchResultTypes'
 import { GenericRecordType } from './GenericRecordTypes';
 import { BroadcastRecordType } from './BroadcastRecordType';
 import { SpellCheckType } from './SpellCheckType';
+import { AxiosRequestConfig } from 'axios';
 
 export interface APISearchResponseType {
 	data: {
@@ -96,4 +97,17 @@ export interface APIAuthResponseType {
 export interface facetItem {
 	name: string;
 	number: number;
+}
+export interface CustomAxiosRequestConfig extends AxiosRequestConfig {
+	_retryCount?: number;
+}
+
+export interface APIAuthMessagesType {
+	data: {
+		streamingBaseUrlVideo: string;
+		AudioUiConfId: string;
+		videoUiConfId: string;
+		partnerId: string;
+		streamingBaseUrlAudio: string;
+	};
 }