+++ src/types/SearchBarTypes.ts
@@ -1,0 +1,4 @@
+export interface PortalSelectItem {
+	name: string;
+	destination: string;
+}